Fix Metadata exchange
We don't expect the bencoding to magically stop at the end of the
structure anymore, we rely on the default metadata piece size or on the
total_size field to deduce the metadata piece size.

+++ peer.go
@@ -284,7 +284,7 @@ func (p *peerState) peerReader(msgChan chan peerMessage) {
 func (p *peerState) sendMetadataRequest(piece int) {
 	metaMessage := MetadataMessage{
 		MsgType: METADATA_REQUEST,
-		Piece:   uint(piece),
+		Piece:   piece,
 	}
 
 	p.sendExtensionMessage("ut_metadata", metaMessage)
